Moved HashtagPanel to separate file
