+++ react/gameday2/components/HashtagPanel.js
@@ -1,0 +1,30 @@
+import React from 'react';
+var classNames = require('classnames');
+
+var HashtagPanel = React.createClass({
+  componentDidMount: function() {
+    !function(d,s,id){
+      var js,fjs=d.getElementsByTagName(s)[0],p=/^http:/.test(d.location)?'http':'https';
+      if(!d.getElementById(id)){
+        js=d.createElement(s);js.id=id;
+        js.src=p+"://platform.twitter.com/widgets.js";
+        fjs.parentNode.insertBefore(js,fjs);
+      }
+    }
+    (document,"script","twitter-wjs");
+  },
+  render: function() {
+    var classes = classNames({
+      'hidden': !this.props.enabled,
+      'pull-left': true,
+      'sidebar': true,
+    });
+    return (
+      <div className={classes}>
+        <div id="twitter-widget"><a className="twitter-timeline" href="https://twitter.com/search?q=%23omgrobots" data-widget-id="406597120632709121">Tweets about "#omgrobots"</a></div>
+      </div>
+    );
+  }
+});
+
+export default HashtagPanel;
